requirements.txt
coverage==4.0.3
defusedxml==0.4.1
dj-database-url==0.4.0
dj-static==0.0.6
Django==1.8.7
django-allauth==0.24.1
django-analytical==2.0.0
django-autocomplete-light==2.2.10
django-axes==1.6.0
django-bootstrap3==7.0.0
django-countries==3.4.1
django-datetime-widget==0.9.3
django-pylibmc==0.6.1
django-toolbelt==0.0.1
djangorestframework==3.3.2
ecdsa==0.13
Fabric==1.10.2
foreman==0.9.7
gunicorn==19.4.5
httplib2==0.9.2
oauth2==1.9.0.post1
oauthlib==1.0.3
paramiko==1.16.0
Pillow==3.0.0
psycopg2==2.6.1
pycrypto==2.6.1
PyJWT==1.4.0
pylibmc==1.5.0
python-openid==2.2.5
python-social-auth==0.2.13
python3-openid==3.0.9
pytz==2015.7
requests==2.8.1
requests-futures==0.9.7
requests-oauthlib==0.5.0
selenium==2.52.0
six==1.10.0
static3==0.6.1
